+++ src/main/java/backend/like_house/domain/chatting/service/impl/ChatCommandServiceImpl.java
@@ -1,9 +1,10 @@
-package backend.like_house.domain.chatting.service;
+package backend.like_house.domain.chatting.service.impl;
 
 import backend.like_house.domain.chatting.converter.ChatConverter;
 import backend.like_house.domain.chatting.entity.ChatRoom;
 import backend.like_house.domain.chatting.repository.ChatRepository;
 import backend.like_house.domain.chatting.repository.ChatRoomRepository;
+import backend.like_house.domain.chatting.service.ChatCommandService;
 import backend.like_house.domain.user.entity.User;
 import backend.like_house.domain.user.repository.UserRepository;
 import backend.like_house.global.error.code.status.ErrorStatus;
@@ -16,7 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class ChatCommandServiceImpl implements ChatCommandService{
+public class ChatCommandServiceImpl implements ChatCommandService {
 
     private final ChatRepository chatRepository;
     private final ChatRoomRepository chatRoomRepository;
