+++ src/main/java/backend/like_house/domain/chatting/service/impl/ChatQueryServiceImpl.java
@@ -1,10 +1,11 @@
-package backend.like_house.domain.chatting.service;
+package backend.like_house.domain.chatting.service.impl;
 
 import backend.like_house.domain.chatting.converter.ChatConverter;
 import backend.like_house.domain.chatting.dto.ChatDTO;
 import backend.like_house.domain.chatting.entity.Chat;
 import backend.like_house.domain.chatting.repository.ChatRepository;
 import backend.like_house.domain.chatting.repository.ChatRoomRepository;
+import backend.like_house.domain.chatting.service.ChatQueryService;
 import backend.like_house.domain.user.entity.User;
 import backend.like_house.global.error.code.status.ErrorStatus;
 import backend.like_house.global.error.handler.ChatRoomException;
